refactor(codegen): extract type-system codegen into types/mod.rs (Round 5)
14 fns lifted into types/mod.rs as inherent impls on Cx (no trait
extension — inherent impls merge across files in Rust, so call sites
don't need any `use` import):

Layout:
- type_expr_to_codegen_ty (the TypeExpr → CodegenTy mapper)
- llvm_basic_type, llvm_array_storage_type
- view_struct_ty, ty_needs_self_field_deep_copy

Composite (user-defined types + generics):
- declare_user_type, populate_user_type_fields
- enum_storage_struct
- resolve_generic_struct_path, resolve_generic_struct_path_for_codegen_ty

Views (F.30):
- wrap_literal_as_view, unpack_view_if_needed

Interface (F.20 vtable dispatch):
- locus_satisfies_interface, coerce_to_interface

Surface bumps to pub(crate):
- Cx fields: pending_locus_names, pending_type_names, user_interfaces
- Cx methods: declare_user_type, resolve_generic_struct_path_for_codegen_ty,
  mangle_generic_name, ensure_vtable
- Free fn: literal_to_view_coerces

This is the smallest extraction sub-round of the refactor (~919 lines
vs Round 4d's 2,675). Type-system codegen turned out more compact
than the plan estimated — much of what could be "types" (interface
vtable synthesis, generic monomorphization) lives inline inside
declare_locus_methods and lower_locus_instantiation (now in
locus/decl.rs + locus/instantiation.rs).

codegen.rs: 28083 -> 27164 (-919 lines).
