+++ crates/hale-codegen/src/lib.rs
@@ -34,6 +34,7 @@ pub mod codegen;
 pub(crate) mod locus;
 pub mod mangle;
 pub(crate) mod stdlib;
+pub(crate) mod types;
 
 pub use codegen::{
     build_executable, build_executable_with_imports,
